only jacobian and hessian missing + sparsity patterns

+++ src/linalg.h
@@ -7,9 +7,9 @@
 // for now its sufficient to memorize where these subroutines are used
 
 namespace Linalg {
-    double dot(const size_t size, const double* x, const double* y);
-    void dsaxpy(const size_t size, const double* x, const double* y, const double* D, double beta, bool invD, double* out);
-    void dgmv(const size_t size, const double* x, const double* y, const double* D, double beta, bool invD, double* out);
+    double dot(const int size, const double* x, const double* y);
+    void dsaxpy(const int size, const double* x, const double* y, const double* D, double beta, bool invD, double* out);
+    void dgmv(const int size, const double* x, const double* y, const double* D, double beta, bool invD, double* out);
 }
 
 #endif  // OPT_LINALG_H
